# ChatBot.Backend/openapi.json
{"openapi":"3.1.0","info":{"title":"FastAPI","version":"0.1.0"},"paths":{"/users/users/":{"post":{"tags":["users"],"summary":"Register User","operationId":"register_user_users_users__post","parameters":[{"name":"init_data","in":"query","required":true,"schema":{"type":"string","title":"Init Data"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects/":{"post":{"tags":["projects"],"summary":"Register Project","operationId":"register_project_api_projects__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["projects"],"summary":"Get Projects","operationId":"get_projects_api_projects__get","parameters":[{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","title":"Telegram Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProjectOut"},"title":"Response Get Projects Api Projects  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/projects/{project_id}/rename":{"patch":{"tags":["projects"],"summary":"Rename Project","operationId":"rename_project_api_projects__project_id__rename_patch","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectRename"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/folders/folders/":{"post":{"tags":["folders"],"summary":"Register Folder","operationId":"register_folder_folders_folders__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FolderCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FolderOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/folders/{folder_id}/rename":{"patch":{"tags":["folders"],"summary":"Rename Folder","operationId":"rename_folder_folders__folder_id__rename_patch","parameters":[{"name":"folder_id","in":"path","required":true,"schema":{"type":"integer","title":"Folder Id"}},{"name":"new_name","in":"query","required":true,"schema":{"type":"string","title":"New Name"}},{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","title":"Telegram Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chats/":{"post":{"tags":["chats"],"summary":"Register Chat","operationId":"register_chat_api_chats__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["chats"],"summary":"Get Chats","description":"Получает чаты пользователя.\n\nArgs:\n    telegram_id: Telegram ID пользователя\n    auth_telegram_id: Telegram ID из заголовка авторизации\n    db: Сессия базы данных\n    api_key: ключ авторизации (для Swagger UI)\n\nReturns:\n    List[ChatOut]: Список чатов пользователя\n\nRaises:\n    HTTPException: Если пользователь не авторизован или не найден","operationId":"get_chats_api_chats__get","security":[{"APIKeyHeader":[]}],"parameters":[{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","description":"Telegram ID пользователя","title":"Telegram Id"},"description":"Telegram ID пользователя"},{"name":"Authorization","in":"header","required":true,"schema":{"type":"string","title":"Authorization"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatOut"},"title":"Response Get Chats Api Chats  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/chats/{chat_id}/rename":{"patch":{"tags":["chats"],"summary":"Rename Chat","operationId":"rename_chat_api_chats__chat_id__rename_patch","parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"integer","title":"Chat Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRename"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/messages/messages/":{"post":{"tags":["messages"],"summary":"Send Message","operationId":"send_message_messages_messages__post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["messages"],"summary":"Get Messages","operationId":"get_messages_messages_messages__get","parameters":[{"name":"chat_id","in":"query","required":true,"schema":{"type":"integer","title":"Chat Id"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":100,"default":50,"title":"Limit"}},{"name":"before_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Before Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageOut"},"title":"Response Get Messages Messages Messages  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/memories/":{"post":{"tags":["memories"],"summary":"Create Memory","operationId":"create_memory_memories__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemoryCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MemoryOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/users/prompt":{"patch":{"tags":["users"],"summary":"Update Custom Prompt","operationId":"update_custom_prompt_users_prompt_patch","parameters":[{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","title":"Telegram Id"}},{"name":"prompt","in":"query","required":true,"schema":{"type":"string","title":"Prompt"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete/chats/{chat_id}":{"delete":{"tags":["delete"],"summary":"Delete Chat","operationId":"delete_chat_delete_chats__chat_id__delete","parameters":[{"name":"chat_id","in":"path","required":true,"schema":{"type":"integer","title":"Chat Id"}},{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","title":"Telegram Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete/projects/{project_id}":{"delete":{"tags":["delete"],"summary":"Delete Project","operationId":"delete_project_delete_projects__project_id__delete","parameters":[{"name":"project_id","in":"path","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","title":"Telegram Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete/folders/{folder_id}":{"delete":{"tags":["delete"],"summary":"Delete Folder","operationId":"delete_folder_delete_folders__folder_id__delete","parameters":[{"name":"folder_id","in":"path","required":true,"schema":{"type":"integer","title":"Folder Id"}},{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","title":"Telegram Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/delete/memory/{memory_id}":{"delete":{"tags":["delete"],"summary":"Delete Memory","operationId":"delete_memory_delete_memory__memory_id__delete","parameters":[{"name":"memory_id","in":"path","required":true,"schema":{"type":"integer","title":"Memory Id"}},{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","title":"Telegram Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/chats/settings":{"patch":{"tags":["chats"],"summary":"Update Chat Settings","operationId":"update_chat_settings_chats_settings_patch","parameters":[{"name":"chat_id","in":"query","required":true,"schema":{"type":"integer","title":"Chat Id"}},{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","title":"Telegram Id"}},{"name":"project_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Project Id"}},{"name":"folder_id","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Folder Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/projects/system-prompt":{"patch":{"tags":["projects"],"summary":"Update Project System Prompt","operationId":"update_project_system_prompt_projects_system_prompt_patch","parameters":[{"name":"project_id","in":"query","required":true,"schema":{"type":"integer","title":"Project Id"}},{"name":"telegram_id","in":"query","required":true,"schema":{"type":"integer","title":"Telegram Id"}},{"name":"system_prompt","in":"query","required":true,"schema":{"type":"string","title":"System Prompt"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/ping":{"get":{"summary":"Ping","operationId":"ping_ping_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ChatCreate":{"properties":{"user_id":{"type":"integer","title":"User Id"},"project_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Project Id"},"folder_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Folder Id"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"model_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Model Id"},"parent_chat_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Chat Id"},"parent_message_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Message Id"}},"type":"object","required":["user_id"],"title":"ChatCreate"},"ChatOut":{"properties":{"user_id":{"type":"integer","title":"User Id"},"project_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Project Id"},"folder_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Folder Id"},"title":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Title"},"model_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Model Id"},"parent_chat_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Chat Id"},"parent_message_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Message Id"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["user_id","id","created_at"],"title":"ChatOut"},"ChatRename":{"properties":{"new_title":{"type":"string","title":"New Title"},"telegram_id":{"type":"integer","title":"Telegram Id"}},"type":"object","required":["new_title","telegram_id"],"title":"ChatRename"},"FolderCreate":{"properties":{"project_id":{"type":"integer","title":"Project Id"},"name":{"type":"string","title":"Name"}},"type":"object","required":["project_id","name"],"title":"FolderCreate"},"FolderOut":{"properties":{"id":{"type":"integer","title":"Id"},"project_id":{"type":"integer","title":"Project Id"},"name":{"type":"string","title":"Name"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","project_id","name","created_at"],"title":"FolderOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MemoryCreate":{"properties":{"user_id":{"type":"integer","title":"User Id"},"content":{"type":"string","title":"Content"}},"type":"object","required":["user_id","content"],"title":"MemoryCreate"},"MemoryOut":{"properties":{"user_id":{"type":"integer","title":"User Id"},"content":{"type":"string","title":"Content"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["user_id","content","id","created_at"],"title":"MemoryOut"},"MessageCreate":{"properties":{"chat_id":{"type":"integer","title":"Chat Id"},"content":{"type":"string","title":"Content"},"role":{"type":"string","title":"Role","default":"user"},"parent_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Id"}},"type":"object","required":["chat_id","content"],"title":"MessageCreate"},"MessageOut":{"properties":{"id":{"type":"integer","title":"Id"},"chat_id":{"type":"integer","title":"Chat Id"},"content":{"type":"string","title":"Content"},"role":{"type":"string","title":"Role"},"parent_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Parent Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","chat_id","content","role","parent_id","created_at"],"title":"MessageOut"},"ProjectCreate":{"properties":{"name":{"type":"string","title":"Name"},"user_id":{"type":"integer","title":"User Id"}},"type":"object","required":["name","user_id"],"title":"ProjectCreate"},"ProjectOut":{"properties":{"name":{"type":"string","title":"Name"},"user_id":{"type":"integer","title":"User Id"},"id":{"type":"integer","title":"Id"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["name","user_id","id","created_at"],"title":"ProjectOut"},"ProjectRename":{"properties":{"new_name":{"type":"string","title":"New Name"},"telegram_id":{"type":"integer","title":"Telegram Id"}},"type":"object","required":["new_name","telegram_id"],"title":"ProjectRename"},"UserOut":{"properties":{"telegram_id":{"type":"integer","title":"Telegram Id"},"username":{"type":"string","title":"Username"},"full_name":{"type":"string","title":"Full Name"},"is_plus":{"type":"boolean","title":"Is Plus"}},"type":"object","required":["telegram_id","username","full_name","is_plus"],"title":"UserOut"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"Authorization"}}}}